// src/app/(frontend)/lib/portfolio/types.ts
export type AboutContent = {
  experience: {
    years: number
    title: string
    description: string
  }
  workPermit: {
    title: string
    description: string
  }
  /** Gallery image URLs. */
  images: string[]
  interests: {
    title: string
    description: string
    areas: string[]
  }
}